# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Remont API
  version: 1.0.0
paths:
...
  /api/v1/repairs/:
    post:
      tags:
      - repairs
      summary: Create Repair
      operationId: create_repair_api_v1_repairs__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepairCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/RepairOutAdmin'
                - $ref: '#/components/schemas/RepairOutUser'
                title: Response Create Repair Api V1 Repairs  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/repairs/{repair_id}:
    get:
      tags:
      - repairs
      summary: Get Repair
      operationId: get_repair_api_v1_repairs__repair_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: repair_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Repair Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/RepairOutAdmin'
                - $ref: '#/components/schemas/RepairOutUser'
                title: Response Get Repair Api V1 Repairs  Repair Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/plans/:
    get:
      tags:
      - plans
      summary: Get Plans
      operationId: get_plans_api_v1_plans__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: machine_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Machine Code
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date To
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanOut'
                title: Response Get Plans Api V1 Plans  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - plans
      summary: Create Plan
      operationId: create_plan_api_v1_plans__post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plans/{plan_id}:
    put:
      tags:
      - plans
      summary: Update Plan
      operationId: update_plan_api_v1_plans__plan_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plans/{plan_id}/complete:
    post:
      tags:
      - plans
      summary: Complete Plan
      operationId: complete_plan_api_v1_plans__plan_id__complete_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/templates/:
    get:
      tags:
      - templates
      summary: Get Templates
      operationId: get_templates_api_v1_templates__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TemplateOut'
                type: array
                title: Response Get Templates Api V1 Templates  Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - templates
      summary: Create Template
      operationId: create_template_api_v1_templates__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/templates/{template_id}:
    get:
      tags:
      - templates
      summary: Get Template
      operationId: get_template_api_v1_templates__template_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - templates
      summary: Patch Template
      operationId: patch_template_api_v1_templates__template_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Template Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplatePatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/templates/{template_id}/copy:
    post:
      tags:
      - templates
      summary: Copy Template
      operationId: copy_template_api_v1_templates__template_id__copy_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PlanCreate:
      properties:
        machine_id:
          type: string
          format: uuid
          title: Machine Id
        type:
          type: string
          title: Type
        plan_date:
          type: string
          format: date-time
          title: Plan Date
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - machine_id
      - type
      - plan_date
      title: PlanCreate
    PlanOut:
      properties:
        machine_id:
          type: string
          format: uuid
          title: Machine Id
        type:
          type: string
          title: Type
        plan_date:
          type: string
          format: date-time
          title: Plan Date
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        id:
          type: string
          format: uuid
          title: Id
        status:
          type: string
          title: Status
        done_repair_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Done Repair Id
      type: object
      required:
      - machine_id
      - type
      - plan_date
      - id
      - status
      title: PlanOut
...
    RepairCreate:
      properties:
        machine_id:
          type: string
          format: uuid
          title: Machine Id
        repair_date:
          type: string
          format: date-time
          title: Repair Date
        type:
          type: string
          title: Type
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        crew:
          anyOf:
          - type: string
          - type: 'null'
          title: Crew
        zaprafka_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Zaprafka Id
        plan_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Plan Id
        items:
          items:
            $ref: '#/components/schemas/RepairItemCreate'
          type: array
          title: Items
      type: object
      required:
      - machine_id
      - repair_date
      - type
      - items
      title: RepairCreate
    RepairItemCreate:
      properties:
        item_no:
          type: integer
          title: Item No
        material_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Material Id
        free_text_material:
          anyOf:
          - type: string
          - type: 'null'
          title: Free Text Material
        condition:
          type: string
          title: Condition
        qty:
          anyOf:
          - type: number
          - type: string
          title: Qty
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - item_no
      - condition
      - qty
      title: RepairItemCreate
    RepairItemOutAdmin:
      properties:
        item_no:
          type: integer
          title: Item No
        material_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Material Id
        free_text_material:
          anyOf:
          - type: string
          - type: 'null'
          title: Free Text Material
        condition:
          type: string
          title: Condition
        qty:
          type: string
          title: Qty
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        id:
          type: string
          format: uuid
          title: Id
        unit_price:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit Price
        amount:
          anyOf:
          - type: string
          - type: 'null'
          title: Amount
      type: object
      required:
      - item_no
      - condition
      - qty
      - id
      title: RepairItemOutAdmin
    RepairItemOutUser:
      properties:
        item_no:
          type: integer
          title: Item No
        material_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Material Id
        free_text_material:
          anyOf:
          - type: string
          - type: 'null'
          title: Free Text Material
        condition:
          type: string
          title: Condition
        qty:
          type: string
          title: Qty
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
      - item_no
      - condition
      - qty
      - id
      title: RepairItemOutUser
    RepairOutAdmin:
      properties:
        machine_id:
          type: string
          format: uuid
          title: Machine Id
        repair_date:
          type: string
          format: date-time
          title: Repair Date
        type:
          type: string
          title: Type
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        crew:
          anyOf:
          - type: string
          - type: 'null'
          title: Crew
        zaprafka_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Zaprafka Id
        plan_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Plan Id
        id:
          type: string
          format: uuid
          title: Id
        status:
          type: string
          title: Status
        version:
          type: integer
          title: Version
        items:
          items:
            $ref: '#/components/schemas/RepairItemOutAdmin'
          type: array
          title: Items
          default: []
      type: object
      required:
      - machine_id
      - repair_date
      - type
      - id
      - status
      - version
      title: RepairOutAdmin
    RepairOutUser:
      properties:
        machine_id:
          type: string
          format: uuid
          title: Machine Id
        repair_date:
          type: string
          format: date-time
          title: Repair Date
        type:
          type: string
          title: Type
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        crew:
          anyOf:
          - type: string
          - type: 'null'
          title: Crew
        zaprafka_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Zaprafka Id
        plan_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Plan Id
        id:
          type: string
          format: uuid
          title: Id
        status:
          type: string
          title: Status
        version:
          type: integer
          title: Version
        items:
          items:
            $ref: '#/components/schemas/RepairItemOutUser'
          type: array
          title: Items
          default: []
      type: object
      required:
      - machine_id
      - repair_date
      - type
      - id
      - status
      - version
      title: RepairOutUser
    TemplateCreate:
      properties:
        name:
          type: string
          title: Name
        type:
          type: string
          title: Type
        items:
          items:
            $ref: '#/components/schemas/TemplateItemCreate'
          type: array
          title: Items
          default: []
      type: object
      required:
      - name
      - type
      title: TemplateCreate
    TemplateItemCreate:
      properties:
        material_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Material Id
        condition:
          type: string
          title: Condition
        qty:
          anyOf:
          - type: number
          - type: string
          title: Qty
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - condition
      - qty
      title: TemplateItemCreate
    TemplateOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        type:
          type: string
          title: Type
        is_active:
          type: boolean
          title: Is Active
      type: object
      required:
      - id
      - name
      - type
      - is_active
      title: TemplateOut
    TemplatePatch:
      properties:
        active:
          type: boolean
          title: Active
      type: object
      required:
      - active
      title: TemplatePatch
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: api/v1/auth/login